layers: Move PV layer data structs into header
